+++ BxInputController/Sources/Common/BxInputCell.swift
@@ -2,7 +2,7 @@
  *	@file BxInputCell.swift
  *	@namespace BxInputController
  *
- *	@details BxInputCell is base cell protocol
+ *	@details BxInputTitleCell is base cell protocol
  *	@date 24.04.2017
  *	@author Sergey Balalaev
  *
@@ -13,17 +13,19 @@
 
 import UIKit
 
-/// BxInputCell is base cell protocol
-public protocol BxInputCell : AnyObject
+/// BxInputTitleCell is base cell protocol with title
+public protocol BxInputTitleCell : AnyObject
 {
     /// caption of the row
     var titleLabel: UILabel! {get}
     /// caption of title for the row. This is only option field
     var subtitleLabel: UILabel? {get}
+    /// for managment of position subtitle
+    var subtitleBottomMargin: NSLayoutConstraint? {get}
 }
 
 /// Base cell protocol with TextField
-public protocol BxInputFieldCell : BxInputCell
+public protocol BxInputFieldCell : BxInputTitleCell
 {
     /// value for the row or a putting example
     var valueTextField: UITextField! {get}