+++ BxInputController/Sources/Common/View/Selector/BxInputSelectorCell.swift
@@ -20,7 +20,7 @@ open class BxInputSelectorCell: UITableViewCell, BxInputFieldCell {
     @IBOutlet public weak var titleLabel: UILabel!
     /// caption of title for the row
     @IBOutlet public weak var subtitleLabel: UILabel?
-    @IBOutlet weak var subtitleBottomMargin: NSLayoutConstraint?
+    @IBOutlet public weak var subtitleBottomMargin: NSLayoutConstraint?
     /// value from selected field
     @IBOutlet public weak var valueTextField: UITextField!
     /// arrow that show state of row closed/opened