+++ BxInputController/Sources/Common/View/Selector/BxInputSelectorRowBinder.swift
@@ -14,7 +14,7 @@
 import UIKit
 
 /// Binder for a selector row
-open class BxInputSelectorRowBinder<Row: BxInputSelectorRow, Cell: BxInputSelectorCell>: BxInputBaseRowBinder<Row, Cell>
+open class BxInputSelectorRowBinder<Row: BxInputSelectorRow, Cell: BxInputSelectorCell>: BxInputBaseFieldRowBinder<Row, Cell>
 {
     /// call when user selected this cell
     override open func didSelected()
@@ -53,39 +53,14 @@ open class BxInputSelectorRowBinder<Row: BxInputSelectorRow, Cell: BxInputSelect
         //
         cell?.arrowImage.image = BxInputUtils.getImage(resourceId: "bx_arrow_to_bottom")
         //
-        cell?.titleLabel.font = owner?.settings.titleFont
-        cell?.titleLabel.textColor = owner?.settings.titleColor
-        cell?.titleLabel.text = row.title
-        if let subtitleLabel = cell?.subtitleLabel {
-            subtitleLabel.font = owner?.settings.subtitleFont
-            subtitleLabel.textColor = owner?.settings.subtitleColor
-            if let margin = owner?.settings.subtitleMargin {
-                cell?.subtitleBottomMargin?.constant = margin
-            }
-            if let alignment = owner?.settings.subtitleAlignment {
-                subtitleLabel.textAlignment = alignment
-            }
-            subtitleLabel.text = row.subtitle
-        }
-        
-        cell?.valueTextField.font = owner?.settings.valueFont
-        cell?.valueTextField.textColor = owner?.settings.valueColor
-        cell?.valueTextField.setPlaceholder(row.placeholder, with: owner?.settings.placeholderColor)
-        
         refreshOpened(animated: false)
-        
         updateChecking()
-        
-        
     }
     
     /// event of change isEnabled
     override open func didSetEnabled(_ value: Bool)
     {
         super.didSetEnabled(value)
-        //cell?.valueTextField.isEnabled = value this is disabled everytime
-        cell?.valueTextField.alpha = value ? 1 : 0.5
-        cell?.titleLabel.alpha = value ? 1 : 0.5
         cell?.arrowImage.alpha = value ? 1 : 0.5
     }